latest version of deciding whether it's a major tracker network or not for the sake of the icon
