+++ shared/js/ui/templates/utils.es6.js
@@ -1,0 +1,6 @@
+const constants = require('./../../../data/constants.js')
+const bkg = chrome.extension.getBackgroundPage()
+
+module.exports = function(trackerName) {
+
+}
